fixes GH-875 - bootstrap yesno button in admin pages which render plugin info

+++ administrator/components/com_fabrik/views/namespace.js
@@ -39,7 +39,7 @@ if (typeof(jQuery) !== 'undefined') {
 				label.closest('.btn-group').find("label").removeClass('active btn-success btn-danger btn-primary');
 				if (input.val() === '') {
 					label.addClass('active btn-primary');
-				} else if (input.val() === 0) {
+				} else if (input.val().toInt() === 0) {
 					label.addClass('active btn-danger');
 				} else {
 					label.addClass('active btn-success');
